simulation tests split to two files and updated
